+++ app/components/sdg_management/relations/index_component.html.erb
@@ -1,15 +1,6 @@
 <%= header %>
 
-<%= render Admin::SearchComponent.new(label: search_label, class: "complex") do |component| %>
-  <%= component.select_tag :goal_code, goal_options,
-                           include_blank: goal_blank_option,
-                           "aria-label": goal_label %>
-  <%= component.select_tag :target_code, target_options,
-                           include_blank: target_blank_option,
-                           "aria-label": target_label %>
-  <%= component.hidden_field_tag :filter, current_filter %>
-<% end %>
-
+<%= render SDGManagement::Relations::SearchComponent.new(label: search_label, current_filter: current_filter) %>
 <%= render "shared/filter_subnav", i18n_namespace: "sdg_management.relations.index" %>
 
 <table>
