+++ app/components/sdg_management/relations/index_component.rb
@@ -1,6 +1,5 @@
 class SDGManagement::Relations::IndexComponent < ApplicationComponent
   include Header
-  include SDG::OptionsForSelect
   delegate :valid_filters, :current_filter, to: :helpers
 
   attr_reader :records
@@ -30,29 +29,5 @@ class SDGManagement::Relations::IndexComponent < ApplicationComponent
 
     def search_label
       t("admin.shared.search.label.#{model_class.table_name}")
-    end
-
-    def goal_label
-      t("admin.shared.search.advanced_filters.sdg_goals.label")
-    end
-
-    def goal_blank_option
-      t("admin.shared.search.advanced_filters.sdg_goals.all")
-    end
-
-    def target_label
-      t("admin.shared.search.advanced_filters.sdg_targets.label")
-    end
-
-    def target_blank_option
-      t("admin.shared.search.advanced_filters.sdg_targets.all")
-    end
-
-    def goal_options
-      super(params[:goal_code])
-    end
-
-    def target_options
-      super(params[:target_code])
     end
 end
