+++ app/components/sdg_management/relations/search_component.html.erb
@@ -1,0 +1,9 @@
+<%= render Admin::SearchComponent.new(label: label, class: "complex") do |component| %>
+  <%= component.select_tag :goal_code, goal_options,
+                           include_blank: goal_blank_option,
+                           "aria-label": goal_label %>
+  <%= component.select_tag :target_code, target_options,
+                           include_blank: target_blank_option,
+                           "aria-label": target_label %>
+  <%= component.hidden_field_tag :filter, current_filter %>
+<% end %>
